moves icon functions to helpers.js
